requirements.txt
Django==3.1.7
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2

djoser==2.1.0
mercadopago==1.2.0
Pillow==8.1.0
PyMySQL==1.0.2

